requirements.txt
argcomplete==1.10.0
beautifulsoup4==4.8.0
chardet==3.0.4
click==7.1.2
cycler==0.10.0
defusedxml==0.6.0
docx2txt==0.8
EbookLib==0.17.1
extract-msg==0.23.1
IMAPClient==2.1.0
joblib==0.14.1
kiwisolver==1.2.0
lxml==4.5.0
matplotlib==3.2.1
nltk==3.5
numpy==1.18.4
odfpy==1.4.1
olefile==0.46
pandas==1.0.3
patool==1.12
pdfminer.six==20181108
Pillow==7.1.2
pycryptodome==3.9.7
pyparsing==2.4.7
PyPDF2==1.26.0
python-dateutil==2.8.1
python-pptx==0.6.18
pytz==2020.1
regex==2020.5.7
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.1
six==1.12.0
sklearn==0.0
sortedcontainers==2.1.0
soupsieve==2.0
SpeechRecognition==3.8.1
textract==1.6.3
tqdm==4.46.0
tzlocal==1.5.1
xlrd==1.2.0
XlsxWriter==1.2.8
